FELIX-6149 : Make webapp support optional
git-svn-id: https://svn.apache.org/repos/asf/felix/trunk@1861331 13f79535-47bb-0310-9956-ffa450edef68
