+++ http/jetty/src/main/java/org/apache/felix/http/jetty/internal/webapp/WebEvent.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.felix.http.jetty.internal;
+package org.apache.felix.http.jetty.internal.webapp;
 
 import java.util.Dictionary;
 import java.util.Hashtable;
